New connection object added to the application object to control the connection check and will handle all the data transfer like battery life  of the other device and more.
Some UI fixes the bug in the background color is fixed.

Some changes on the Threads logic some issue with detecting connection lost and handle streaming. Added Actions to the TCP Connection and a listener for that(Mainly for the Ongoing Notification).

App now detect connection lost even when not streaming with a few seconds delay.

new server and client button design.(new png and selector).

user now need to select the port they want to open.

some not relevant classes is added and will be removed on the next commit.

App now stop the server socket accept in real time(prev was closing after the connection is made.)